pyproject.toml
[project]
name = "src"
version = "1.0.0"
description = ""
authors = [
    {name = "macalistervadim",email = "macalistervadim@yandex.ru"}
]

requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-nested-admin (>=4.1.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "drf-spectacular (==0.27.0)",
]

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"
mypy = "^1.15.0"
ruff = "^0.9.5"
django-stubs = "^5.1.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    "*/migrations/*",
    "venv/",
    "__init__.py",
    "apps.py",
]
line-length = 79
indent-width = 4

[tool.ruff.lint]
select = ["E", "W", "C", "F", "N", "COM", "I", "Q", "U", "N", "A", "M"]
ignore = ["F405"]
fixable = ["ALL"]
unfixable = []



